rename existing package and add a few new ones
