+++ src/main/java/com/management/store/app/ManagementAppApplication.java
@@ -1,4 +1,4 @@
-package com.management.store.ManagementApp;
+package com.management.store.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class ManagementAppApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(ManagementAppApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(ManagementAppApplication.class, args);
+    }
 
 }
