+++ src/test/java/com/management/store/app/ManagementAppApplicationTests.java
@@ -1,4 +1,4 @@
-package com.management.store.ManagementApp;
+package com.management.store.app;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
